Move client protocol specific make_queue functions out of player.c/h
